[QDP] make dataflow async and improve amplitudeEncoder  (#675)
* async pipline

* prevent error

* Add TODO for dynamic chunk size tuning

Added a TODO comment to tune chunk size dynamically based on GPU model or PCIe bandwidth.

* update

* add a test

+++ qdp/qdp-kernels/src/amplitude.cu
@@ -39,8 +39,8 @@ extern "C" {
 int launch_amplitude_encode(
     const double* input_d,
     void* state_d,
-    int input_len,
-    int state_len,
+    size_t input_len,
+    size_t state_len,
     double norm,
     cudaStream_t stream
 ) {
